// _deprecated/src/database/typings/Models.d.ts
export interface Models {
  BomCapsulation: typeof BomCapsulation;
  BoMDataBible: typeof BoMDataBible;
  BomConnection: typeof BomConnection;
  BomDivision: typeof BomDivision;
  BomIndex: typeof BomIndex;
  BomLabel: typeof BomLabel;
  BomLog: typeof BomLog;
  BomLookup: typeof BomLookup;
  BomMapMove: typeof BomMapMove;
  BomMapStory: typeof BomMapStory;
  BomMapMovePeople: typeof BomMapMovePeople;
  BomMapMoveCoords: typeof BomMapMoveCoords;
  BomMap: typeof BomMap;
  BomNarration: typeof BomNarration;
  BomMarkdown: typeof BomMarkdown;
  BomPage: typeof BomPage;
  BomPeopleRels: typeof BomPeopleRels;
  BomPeople: typeof BomPeople;
  BomPlacesCoords: typeof BomPlacesCoords;
  BomPlaces: typeof BomPlaces;
  BomObjects: typeof BomObjects;
  BomXrels: typeof BomXrels;
  BomQuote: typeof BomQuote;
  BomReferences: typeof BomReferences;
  BomSection: typeof BomSection;
  BomSectionrow: typeof BomSectionrow;
  BomShortlinks: typeof BomShortlinks;
  BomSlug: typeof BomSlug;
  BomText: typeof BomText;
  BomTimeline: typeof BomTimeline;
  BomTranslationLocked: typeof BomTranslationLocked;
  BomTranslation: typeof BomTranslation;
  BomUserSocial: typeof BomUserSocial;
  BomUserToken: typeof BomUserToken;
  BomUser: typeof BomUser;
  BomXtrasSaractorsSim: typeof BomXtrasSaractorsSim;
  BomXtrasCaractors: typeof BomXtrasCaractors;
  BomXtrasCommentary: typeof BomXtrasCommentary;
  BomXtrasFax: typeof BomXtrasFax;
  BomXtrasFaxIndex: typeof BomXtrasFaxIndex;
  BomXtrasImage: typeof BomXtrasImage;
  BomXtrasHistory: typeof BomXtrasHistory;
  BomXtrasChiasmus: typeof BomXtrasChiasmus;
  BomXtrasSource: typeof BomXtrasSource;
  BomXtrasStats: typeof BomXtrasStats;
  LdsScripturesBooks: typeof LdsScripturesBooks;
  LdsScripturesFootnotes: typeof LdsScripturesFootnotes;
  LdsScripturesTranslations: typeof LdsScripturesTranslations;
  LdsScripturesLines: typeof LdsScripturesLines;
  LdsScripturesMeta: typeof LdsScripturesMeta;
  LdsScripturesVerses: typeof LdsScripturesVerses;
  LdsScripturesVolumes: typeof LdsScripturesVolumes;
  SocialPosts: typeof SocialPosts;
  
  // Messenger models (Sendbird replacement - Phase 1)
  MessengerUser: typeof MessengerUser;
  MessengerChannel: typeof MessengerChannel;
  MessengerMember: typeof MessengerMember;
  MessengerMessage: typeof MessengerMessage;
  MessengerHighlight: typeof MessengerHighlight;
  MessengerReaction: typeof MessengerReaction;
  MessengerFile: typeof MessengerFile;
  
  [key: string]: typeof ModelBase;
}